requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0